Add a new function to get private alliance hash.
We need to set a password for the to-be-created or already existing MUC
channel, but it needs to be based on something which is only available
for alliance members.

After some research, I finally found something that could be useful for
that (JoinStep). So, if the first member of the alliance should leave,
the channel isn't accessible anymore, so everyone needs to get out of
the channel and rejoin so that the new password is based on the now
oldest member.