Implement listener for the basebuilder button.
Yep, now we are coming to an end, because this branch as well as the chat_window
branch depend on the userscript_compiler branch.

Not necessarily a userscript compiler is the solution for this particular
problem, it's just the nicest one.

Basically we need the ability to attach to certain stages of the script loading
chain of the current page (and thus of the game), as custom windows need to
access the webinterface namespace, which is loaded last.