Add GetReady which generates the loader template.
This is currently a bit of a hack, but it gets its job done, for a clean
solution there still is the userscript compiler in the works.

It currently just wraps the code inside the generate() method which in turn is
expected to write out the whole loader template plus the code itself.

At the moment, the code is not escaped, so we're generating invalid JS here.