Implement preliminary watch for webfrontend.
This is not working at the moment, as we really need to inject _twice_. But
do so a few commits later.

It is implemented as a timer using window.setInterval() and every function has
the whole body of the userscript itself. So _THAT_ could definitely become a big
memory hog on the *client-side*.

We might want to avoid that (if possible), but for testing its okay.