Pass compiled code to the template constructor.
The UserScriptTemplate now eats code in its constructor and thus doesn't need to
fiddle everything together in the generate() method.

This also means, that the template class itself is responsible for escaping the
code, which could change in some future version, as we might want to have this
within the UserScript library (among other injection features).