Finish style of basebuilder widget.
Okay, this seriously still looks ugly as hell, but for the moment, it provides
the functionality and doesn't overlap other content, so let's go with it for
now.

The basebuilder window currently is just a window showing the URL, but as soon
as we have implemented more external bindings to webfrontend and qx, our own
interfaces will become more sophisticated.