Preliminary implementation of basebuilder IDs.
This currently works well for all buildings, except for harvesters, as we can't
tell from the buildings info whether it's a tiberium harvester or a crystal
harvester. But for the rest it should work fine, though the conversion list of
C&C TA building types to BaseBuilder's building types isn't complete.