Recognize resource fields without harvesters.
We are using City.GetResourceType() in order to prepopulate the map array with
resource fields. Afterwards we just need to check if the building is a harvester
or not. No need to do fuzzy things with get_ProductionModifierTypeDBId().

But let's keep the external binding for this method, we don't know when we will
need it at some later point. And besides, having a more and more complete API of
the game is also good for documentation purposes.