Retrieve timestamps from chat messages.
Delayed timestamps from the server are in UTC, so we need to add the UTC time
zone offset. This is somewhat risky, as we are not yet sure if all browsers will
support this. But even when not, they're delayed stamps, so who cares?

If anyone has a more portable solution, don't hesitate to create a pull request.