Improve the way chat lines are added.
We were using one single Label before, where text was appended by overriding the
value. That way we had to do bookkeeping of two separate variables for the
content and to insert sluggish separators to cut the backbuffer/history.

This now is much much cleaner, as we just use a simple VBox which is filled with
a Label on each line. This is also much cheaper in content.