Add/remove members on groupchat enter/leave.
Which means, that it is in a working state right now, so we can see in realtime
when members leave and/or change nicks.

In addition this fixes a small issue, which caused that nickname changes were
seen as "joins" (for the old nick) as XMPP is sending a 303 with the _active_
presence.