Note in output file that content is autogenerated.
Unfortunately this is necessary because we had the first case where someone
changed the generated userscript and put it online with an incremented version
and the same namespace.

Though this isn't particularly bad from the point of modifying alone (hey, it's
open source, so we strongly encourage modifications), but modifying the real
source code can be done in much shorter time, as you don't need to dig through
minified JavaScript.

In addition, people contributing directly to the corresponding open source
project will also have a better chance of getting their changes included into
the upstream version.